Reimplement how search bar works
Searchbar was previously a cell that was rendered in the collection view. This had caused multiple issues e.g. keyboard being dismissed while typing, keyboard styles being reset while styping. More prominent in Location Search.

It has been reimplemented to be a sibling of the collection view instead. This allows us to reload certain part of the search bar and/or filter options while keeping the keyboard presentation intact.